requirements.txt
cachetools~=5.3.0
Flask~=2.2.3
SQLAlchemy==1.4.47
requests~=2.28.2
python-dotenv~=1.0.0
pydantic~=1.10.7